v2.0: Merge PageXML and Page class (only one page possible) and improved class names.
