+++ pyproject.toml
@@ -6,9 +6,9 @@ build-backend = "setuptools.build_meta"
 name = "pypxml"
 description = "A python library for parsing, converting and modifying PageXML files. "
 keywords = ["PageXML", "XML", "OCR", "optical character recognition"]
-version = "1.0"
+version = "2.0"
 readme = "README.md"
-license = { file = "LICENSE" }
+license = { text = "MIT License" }
 authors = [
     { name="Janik Haitz", email="jahtz.dev@proton.me" },
 ]
