+++ src/pypxml/__init__.py
@@ -2,8 +2,7 @@
 # Copyright (c) 2024 Janik Haitz
 # See the LICENSE file in the root directory for more details.
 
-from .pxml import PageXML
-from .page import Page
-from .element import Element
-from .resources.xml_schema import XMLSchema
-from .resources.xml_types import XMLType
+from .page_xml import PageXML
+from .page_element import PageElement
+from .page_types import PageType
+from .page_schema import PageSchema
