+++ src/pypxml/page_element.py
@@ -6,19 +6,23 @@ from typing import Optional, Union, Self
 
 from lxml import etree
 
-from .resources.xml_types import XMLType
+from .page_types import PageType
 
 
-class Element:
-    """ Represents an element inside a page. """
-    def __init__(self, _type: XMLType, attributes: Optional[dict[str, str]] = None):
-        self.__type: XMLType = _type
+class PageElement:
+    def __init__(self, _type: PageType, **attributes: str) -> None:
+        """
+        Please use the .new() constructor.
+        :param _type: PageType of this PageElement.
+        :param attributes: Attributes of this PageElement.
+        """
+        self.__type: PageType = _type
         self.__attributes: dict[str, str] = attributes if attributes else {}
-        self.__elements: list[Element] = []
+        self.__elements: list[PageElement] = []
         self.__text: Optional[str] = None
 
-    def __len__(self) -> int:
-        """ Returns the number of elements. """
+    def __len__(self):
+        """ Returns the number of sub elements. """
         return len(self.__elements)
 
     def __iter__(self) -> Self:
@@ -36,9 +40,9 @@ class Element:
 
     def __getitem__(self, key: Union[int, str]) -> Optional[Union[Self, str]]:
         """
-        Get an Element object by its index or an attribute value by its key
-        :param key: Index (integer) of an Element object or a key (string) of an attribute.
-        :return: The Element of passed index (returns last object if the key is out of range) or the value of the
+        Get an PagElement object by its index or an attribute value by its key
+        :param key: Index (integer) of an PagElement object or a key (string) of an attribute.
+        :return: The PagElement of passed index (returns last object if the key is out of range) or the value of the
             selected attribute. Returns None, if no match was found.
         """
         if isinstance(key, int) and len(self.__elements) > 0:
@@ -49,11 +53,11 @@ class Element:
 
     def __setitem__(self, key: Union[int, str], value: Union[Self, str]) -> None:
         """
-        Set an Element object or an attribute value.
-        :param key: Index (integer) for an Element object or a key (string) for an attribute.
-        :param value: Element object (if key is of type integer) or a string (if key is of type string).
+        Set an PagElement object or an attribute value.
+        :param key: Index (integer) for an PagElement object or a key (string) for an attribute.
+        :param value: PagElement object (if key is of type integer) or a string (if key is of type string).
         """
-        if isinstance(key, int) and isinstance(value, Element) and len(self.__elements) > 0:
+        if isinstance(key, int) and isinstance(value, PageElement) and len(self.__elements) > 0:
             self.__elements[min(key, len(self.__elements) - 1)] = value
         elif isinstance(key, str):
             self.__attributes[key] = value
@@ -62,38 +66,44 @@ class Element:
 
     def __contains__(self, key: Union[Self, str]) -> bool:
         """
-        Checks if an Element object or an attribute exists.
-        :param key: Element object or attribute key.
-        :return: True, if either the passed Element object or the attribute exists. Else return False.
+        Checks if an PagElement object or an attribute exists.
+        :param key: PagElement object or attribute key.
+        :return: True, if either the passed PagElement object or the attribute exists. Else return False.
         """
-        if isinstance(key, Element):
+        if isinstance(key, PageElement):
             return key in self.__elements
         elif isinstance(key, str):
             return key in self.__attributes
         return False
 
     @property
-    def type(self) -> XMLType:
+    def type(self) -> PageType:
+        """ Type of this PageElement object. """
         return self.__type
 
     @type.setter
-    def type(self, value: XMLType) -> None:
+    def type(self, value: PageType) -> None:
+        """ Type of this PageElement object. """
         self.__type = value
 
     @property
     def attributes(self) -> dict[str, str]:
+        """ List of all attributes. """
         return self.__attributes
 
     @property
     def elements(self) -> list[Self]:
+        """ List of all elements. """
         return self.__elements
 
     @property
     def id(self) -> Optional[str]:
+        """ ID attribute. """
         return self.__attributes.get('id', None)
 
     @id.setter
     def id(self, value: Optional[str]) -> None:
+        """ ID attribute. """
         if value is None:
             self.__attributes.pop('id', None)
         else:
@@ -101,40 +111,42 @@ class Element:
 
     @property
     def text(self) -> Optional[str]:
+        """ XML element text. """
         return self.__text
 
     @text.setter
     def text(self, value: Optional[str]) -> None:
+        """ XML element text. """
         self.__text = None if value is None else str(value)
 
     @classmethod
-    def new(cls, _type: XMLType, **attributes: str) -> Self:
+    def new(cls, _type: PageType, **attributes: str) -> Self:
         """
-        Create a new Element object from scratch.
-        :param _type: The type of element to create.
+        Create a new PageElement object from scratch.
+        :param _type: The type of page element to create.
         :param attributes: Named arguments that will be stores as xml attributes.
-        :return: The newly created Element object.
+        :return: The newly created PageElement object.
         """
         attributes = {str(k): str(v) for k, v in attributes.items() if v is not None}
-        return cls(_type, attributes)
+        return cls(_type, **attributes)
 
     @classmethod
     def from_etree(cls, tree: etree.Element) -> Self:
         """
-        Create a new Element object from a lxml etree object.
+        Create a new PageElement object from a lxml etree object.
         :param tree: lxml etree object.
-        :return: Element object that represents the passed etree object.
+        :return: PageElement object that represents the passed etree object.
         """
-        element = cls(XMLType(tree.tag.split('}')[1]), dict(tree.items()))
+        element = cls(PageType(tree.tag.split('}')[1]), **dict(tree.items()))
         element.text = tree.text
         for child in tree:
-            element.add_element(Element.from_etree(child))
+            element.add_element(PageElement.from_etree(child))
         return element
 
     def to_etree(self) -> etree.Element:
         """
-        Convert the Element object to a lxml etree object.
-        :return: A lxml etree object that represents this Element object.
+        Convert the PageElement object to a lxml etree object.
+        :return: A lxml etree object that represents this PageElement object.
         """
         element = etree.Element(self.__type.value, **self.__attributes)
         if self.__text is not None:
@@ -147,9 +159,12 @@ class Element:
         """ Returns True, if the Element object is a region. """
         return self.__type.value.endswith('Region')
 
-    def contains_text(self) -> bool:
-        """ Returns True, if the Element object contains text. """
-        return self.__text is not None
+    def get_attribute(self, key: str) -> Optional[str]:
+        """
+        Get an attribute.
+        :param key: Key of attribute.
+        """
+        return self.__attributes.get(key, None)
 
     def set_attribute(self, key: str, value: Optional[str]) -> None:
         """
@@ -171,22 +186,22 @@ class Element:
         return self.__attributes.pop(str(key), None)
 
     def get_coords(self) -> Optional[Self]:
-        """ Return the first direct child Element object of type Coords. """
+        """ Return the first direct child PageElement object of type Coords. """
         for element in self.__elements:
-            if element.type == XMLType.Coords:
+            if element.type == PageType.Coords:
                 return element
         return None
 
     def get_baseline(self) -> Optional[Self]:
-        """ Return the first direct child Element object of type Baseline. """
+        """ Return the first direct child PageElement object of type Baseline. """
         for element in self.__elements:
-            if element.type == XMLType.Baseline:
+            if element.type == PageType.Baseline:
                 return element
         return None
 
     def add_element(self, element: Self, index: Optional[int] = None) -> None:
         """
-        Add an existing Element object to the list elements.
+        Add an existing PageElement object to the list elements.
         :param element: The element to add.
         :param index: If set, insert the element at this index. Else append to the list.
         """
@@ -195,31 +210,31 @@ class Element:
         else:
             self.__elements.insert(min(index, len(self.__elements) - 1), element)
 
-    def create_element(self, _type: XMLType, index: int = None, **attributes: str) -> Self:
+    def create_element(self, _type: PageType, index: int = None, **attributes: str) -> Self:
         """
-        Create a new Element object and add it to the list of elements.
-        :param _type: XMLType of new element.
+        Create a new PageElement object and add it to the list of elements.
+        :param _type: PageType of new element.
         :param index: If set, insert the new element at this index. Else append to the list.
         :param attributes: Named arguments that will be stores as xml attributes.
-        :return: The newly created Element object.
+        :return: The newly created PageElement object.
         """
-        element = Element.new(_type, **attributes)
+        element = PageElement.new(_type, **attributes)
         self.add_element(element, index)
         return element
 
     def remove_element(self, element: Union[Self, int]) -> Optional[Self]:
         """
         Remove an element from the list of elements.
-        :param element: The Element object or the index of the element to remove.
+        :param element: The PagElement object or the index of the element to remove.
         :return: The removed element, if it existed.
         """
         if isinstance(element, int) and element < len(self.__elements) - 1:
             return self.__elements.pop(element)
-        elif isinstance(element, Element) and element in self.__elements:
+        elif isinstance(element, PageElement) and element in self.__elements:
             self.__elements.remove(element)
             return element
         return None
 
     def clear_elements(self) -> None:
-        """ Remove all Element objects from the list of elements. """
+        """ Remove all PagElement objects from the list of elements. """
         self.__elements.clear()
