--- src/pypxml/resources/__init__.py
@@ -1,4 +1,0 @@
-# This file is licensed under the MIT License.
-# Copyright (c) 2024 Janik Haitz
-# See the LICENSE file in the root directory for more details.
-
